requirements.txt
protobuf ~= 3.20
grpcio-tools == 1.30
flask ~= 1.1
Jinja2 ~= 2.11
pytest ~= 5.4